refactor: Simplify and speed up attribute handling
