+++ components/TableOfContents.json
@@ -21,9 +21,10 @@
                 {
                   "element": "a",
                   "attributes": {
-                    "x-href": "state.value.parentElement?.attributes.href.value",
                     "x": "state.value.textContent",
-                    "x-class": "[state.value.textContent === parent.closest.textContent && 'underline', state.tagName === 'H3' && 'ml-2']"
+                    "x-attr": "",
+                    "@href": "state.value.parentElement?.attributes.href.value",
+                    "@class": "[state.value.textContent === parent.closest.textContent && 'underline', state.tagName === 'H3' && 'ml-2']"
                   }
                 }
               ]
